Camera tilemap culling mostly working
Some strange issues occur where tiles on the right will not be visible until the camera is moved further down. When its moved back up, they're no longer rendered. Likely an issue in my calculations. Tilemap rendering code was isolated to the class as it is not an entity in the context of this project. Set screen resolution to 1280x720.